requirements-ml.txt
# Dependencies for the ML pipeline portion (classifier training, offline
# ingestion orchestrator, Vertex AI deploy). Kept separate from the root
# requirements.txt to avoid the existing merge conflict on that file and to
# keep the API container image lean.
#
# Boston's API container does NOT need these.
# Vertex AI prediction container uses deploy/requirements-vertex.txt instead.

# Classifier training
scikit-learn>=1.4
joblib>=1.3
mlflow>=2.10

# Offline ingestion (relies on Boston's api.document_processing + api.vector_store)
openai>=1.30
python-dotenv>=1.0
chromadb>=0.5,<0.6

# Vertex AI deploy
google-cloud-aiplatform>=1.60
PyYAML>=6.0

# Dev / CI
pytest>=8.0
ruff>=0.5
